+++ BackEnd/Order/src/main/java/com/example/order/Client/Inventory/Dto/Response/InventoryWarehouseResponse.java
@@ -12,6 +12,7 @@ import lombok.NoArgsConstructor;
 import lombok.experimental.FieldDefaults;
 import lombok.experimental.SuperBuilder;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 
@@ -26,7 +27,7 @@ public class InventoryWarehouseResponse {
     String product;
     String warehouse;
     String bin;
-    Integer quantity;
+    BigDecimal quantity;
     LocalDate expiryDate;
     String status;
     LocalDateTime createdAt;
